Use Redis as a server side session store
On the staging service on Heroku, we were encountering ActionDispatch::Cookies::CookieOverflow
errors on attempting to log into Auth0. This is a known issue related to the
size of the cookie store in Rails (4kb).

In other dxw projects we have used Redis as a server side session store. The
Redis dyno is available on Heroku and simple to set up (it has been provisioned
on our staging Heroku before this PR was submitted).